Ch8/tesseract/requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
fastapi==0.72.0
h11==0.13.0
idna==3.3
numpy==1.22.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
Pillow==9.0.0
pydantic==1.9.0
pytesseract==0.3.8
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.0
